+++ praw/internal.py
@@ -19,16 +19,15 @@ The functions in this module are not to be relied upon by third-parties.
 """
 
 from __future__ import print_function, unicode_literals
-
-from requests import Request, codes
 import re
 import six
 import sys
+from requests import Request, codes, exceptions
 from requests.compat import urljoin
 from praw.decorators import restrict_access
-from praw.errors import (InvalidSubreddit, NotFound, OAuthException,
-                         OAuthInsufficientScope, OAuthInvalidToken,
-                         RedirectException)
+from praw.errors import (HTTPException, Forbidden, NotFound, InvalidSubreddit,
+                         OAuthException, OAuthInsufficientScope,
+                         OAuthInvalidToken, RedirectException)
 
 
 RE_REDIRECT = re.compile('(rand(om|nsfw))|about/sticky')
@@ -180,9 +179,16 @@ def _raise_response_exceptions(response):
             raise OAuthInvalidToken('invalid_token', response.url)
         else:
             raise OAuthException(msg, response.url)
-    if response.status_code == codes.not_found:
-        raise NotFound(response)
-    response.raise_for_status()  # TODO: Map all codes to a PRAWException
+
+    if response.status_code == codes.forbidden:  # pylint: disable=E1101
+        raise Forbidden(_raw=response)
+    elif response.status_code == codes.not_found:  # pylint: disable=E1101
+        raise NotFound(_raw=response)
+    else:
+        try:
+            response.raise_for_status()  # These should all be directly mapped
+        except exceptions.HTTPError as exc:
+            raise HTTPException(_raw=exc.response)
 
 
 def _to_reddit_list(arg):